refactor(scheduler): split tests into per-domain sibling files
Convert the in-file `mod tests` (665 lines, all of `scheduler/mod.rs`
below the production code) into a `scheduler/tests/` directory module
following the same pattern we used for `orchestrator/tests/`:

  crates/runtime/src/scheduler/tests/
  ├── mod.rs           shared fixtures (build, build_with_agent_dir,
  │                    mount_answer, ollama_answer, AgentDirGuard) +
  │                    module declarations
  ├── cron.rs          compute_next_run (3 tests)
  ├── dispatch.rs      run_due_tasks dispatch + once/recurring (4 tests)
  ├── heartbeat.rs     run_heartbeat (3 tests)
  ├── home_channel.rs  resolve_home_channel_tools + empty fallback (3 tests)
  ├── identity.rs      persona-owner identity threading (1 test)
  └── reap.rs          reap_stale_and_recover (4 tests)

`scheduler/mod.rs` drops from 889 → 215 lines (just the loop machinery
+ trait/built-in tasks). `compute_next_run` and
`resolve_home_channel_tools` are now `#[cfg(test)]`-gated re-exports
since they're only reached by tests today.

cargo test -p assistant-runtime --lib scheduler:: reports the same
18 passing; cargo test -p assistant-runtime --lib still 220 passing.
